\name{gpredictr-package}
\alias{gpredictr-package}
\alias{gpredictr}
\docType{package}
\title{
R Wrapper for Google Prediction API
}
\description{
Provides an R interface to the Google Prediction API v0.15. Includes connection
handlers for insert (formerly "train"), get ("check"), predict, list, and auth.
Includes a result handler for returned JSON objects.
}
\details{
  \tabular{ll}{
  Package: \tab gpredictr\cr
  Type: \tab Package\cr
  Version: \tab 1.0\cr
  Date: \tab 2013-05-04\cr
  License: \tab Apache 2.0\cr
  }
}
\author{
Authors:  Cheng-Neng (Markko) Ko, Maciej Janiec, Victor W. Yee

Maintainer: Victor W. Yee <w.victoryee@gmail.com>

gpredictr is forked from the googlepredictionapi library by Maciej Janiec
(http://code.google.com/p/r-google-prediction-api-v12/,
which is itself a fork from the googlepredictionapi library by Markko Ko
<markko@google.com> at https://code.google.com/p/google-prediction-api-r-client/.
}
\references{
  \enumerate{
   \item https://developers.google.com/prediction/docs/reference/v1.5/
  }
}